requirements.txt
#requirements.txt
aenum==3.1.15
aiofiles==22.1.0
aiohttp==3.8.6
aiosignal==1.3.1
aiostream==0.5.2
alembic==1.13.2
annotated-types==0.7.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.2.0
backoff==2.2.1
backports.tarfile==1.2.0
bcrypt==4.1.3
beautifulsoup4==4.11.2
blinker==1.6.2
Brotli==1.1.0
cachetools==5.3.3
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.6
dataclasses-json==0.5.14
Deprecated==1.2.14
distro==1.9.0
dnspython==2.6.1
docx2txt==0.8
e2b==0.10.6
exceptiongroup==1.2.1
Faker==26.0.0
ffmpeg==1.4
filelock==3.15.4
Flask==3.0.0
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-Testing==0.8.1
frozenlist==1.4.1
fsspec==2024.6.1
google-api-core==2.19.1
google-api-python-client==2.85.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.63.2
gpt3discord==12.3.8
greenlet==3.0.3
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.23.4
idna==3.7
iniconfig==2.0.0
itsdangerous==2.1.2
jaraco.context==5.3.0
Jinja2==3.1.2
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonrpcclient==4.0.3
langchain==0.0.335
langsmith==0.0.92
llama-index==0.8.69.post1
loguru==0.7.2
lxml==5.2.2
Mako==1.3.5
MarkupSafe==2.1.3
marshmallow==3.21.3
more-itertools==10.3.0
multidict==6.0.5
mutagen==1.47.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
nltk==3.8.1
numpy==1.26.4
openai==1.2.0
packaging==23.2
pandas==2.2.2
passlib==1.7.4
Pillow==9.3.0
pinecone-client==2.1.0
pluggy==1.5.0
proto-plus==1.24.0
protobuf==3.20.2
py-cord==2.4.1
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycord-multicog==1.0.2
pycryptodomex==3.20.0
pydantic==2.4.2
pydantic_core==2.10.1
PyJWT==2.8.0
pyparsing==3.1.2
pypdf==3.11.1
pytest==7.2.2
pytest-asyncio==0.21.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-pptx==0.6.21
pytz==2024.1
PyYAML==6.0.1
regex==2024.5.15
replicate==0.15.4
requests==2.31.0
rsa==4.9
sentencepiece==0.1.99
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.31
sqlitedict==2.1.0
tabulate==0.9.0
tenacity==8.5.0
tiktoken==0.4.0
tokenizers==0.13.3
tomli==2.0.1
tqdm==4.64.1
transformers==4.25.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
Unidecode==1.3.6
uritemplate==4.1.1
urllib3==1.26.19
websockets==11.0.3
Werkzeug==3.0.1
wolframalpha==5.0.0
wrapt==1.16.0
XlsxWriter==3.2.0
xmltodict==0.13.0
yarl==1.9.4
youtube-transcript-api==0.5.0
yt-dlp==2023.3.4

